move plugin to flytekit core
