--- plugins/flytekit-inference/flytekitplugins/inference/__init__.py
@@ -1,13 +1,0 @@
-"""
-.. currentmodule:: flytekitplugins.inference
-
-.. autosummary::
-   :template: custom.rst
-   :toctree: generated/
-
-   NIM
-   ModelInferenceTemplate
-"""
-
-from .nim.serve import NIM
-from .sidecar_template import ModelInferenceTemplate
