--- plugins/flytekit-inference/setup.py
@@ -1,38 +1,0 @@
-from setuptools import setup
-
-PLUGIN_NAME = "inference"
-
-microlib_name = f"flytekitplugins-{PLUGIN_NAME}"
-
-plugin_requires = ["flytekit>=1.12.2,<2.0.0", "kubernetes", "openai"]
-
-__version__ = "0.0.0+develop"
-
-setup(
-    name=microlib_name,
-    version=__version__,
-    author="flyteorg",
-    author_email="admin@flyte.org",
-    description="This package enables seamless use of model inference sidecar services within Flyte",
-    namespace_packages=["flytekitplugins"],
-    packages=[f"flytekitplugins.{PLUGIN_NAME}", f"flytekitplugins.{PLUGIN_NAME}.nim"],
-    install_requires=plugin_requires,
-    license="apache2",
-    python_requires=">=3.8",
-    classifiers=[
-        "Intended Audience :: Science/Research",
-        "Intended Audience :: Developers",
-        "License :: OSI Approved :: Apache Software License",
-        "Programming Language :: Python :: 3.8",
-        "Programming Language :: Python :: 3.9",
-        "Programming Language :: Python :: 3.10",
-        "Programming Language :: Python :: 3.11",
-        "Programming Language :: Python :: 3.12",
-        "Topic :: Scientific/Engineering",
-        "Topic :: Scientific/Engineering :: Artificial Intelligence",
-        "Topic :: Software Development",
-        "Topic :: Software Development :: Libraries",
-        "Topic :: Software Development :: Libraries :: Python Modules",
-    ],
-    entry_points={"flytekit.plugins": [f"{PLUGIN_NAME}=flytekitplugins.{PLUGIN_NAME}"]},
-)
